GA-46 created sub folders for the server and client. added READMEs
