+++ .dockerignore
@@ -1,1 +1,1 @@
-build/
+server_cpp/build/
