+++ README.md
@@ -1,47 +1,42 @@
 # Gmail-AdvancedSystemProgramming
-# Dear TA please check main1-Ex1 for the final version of a specific assigment
-## Set up the app
-```bash
-git clone https://github.com/YuvalAnteby/Gmail-AdvancedSystemProgramming.git
-cd Gmail-AdvancedSystemProgramming
-```
-1. build
-```bash
-docker-compose build app
-```
-2. run
-```bash
-docker-compose run --rm app
-```
-3. Deleting manually the .txt files (in terminal)
-```bash
-docker run --rm -v gmail-advancedsystemprogramming_app_data:/data alpine sh -c "rm -f /data/build/data/*.txt"
-```
 
-### (Reminder) Exiting the docker container
-```bash
-control + d
-```
-or 
-```bash
-control + c
-```
+## Dear TA please check main1-Ex1 for the final version of a specific assigment
+## Dear TA please check main1-Ex2 for the final version of a specific assigment
 
-### Rebuilding the code in case of code errors (deletes .txt files)
-```bash
-docker-compose down -v
-docker-compose build app
-docker-compose run --rm app
-```
+---
 
-### Testing locally with Docker
-```bash
-git clone https://github.com/YuvalAnteby/Gmail-AdvancedSystemProgramming.git
-cd Gmail-AdvancedSystemProgramming
-docker-compose run --build --rm gtest
-```
-## Screenshots
-<img src="screenshots/BloomFilter/example run 1.png" height="500" alt="img1">
-<img src="screenshots/BloomFilter/example run 2.png" height="500" alt="img2">
-<img src="screenshots/BloomFilter/example run 3.png" height="500" alt="img3">
-<img src="screenshots/BloomFilter/test results.png" height="500" alt="img4">
+## Table of Contents
+- [Screenshots](#screenshots)
+  - [Ex1](#ex1-screenshots)
+  - [Ex2](#ex2-screenshots)
+- [Useful links](#useful-links)
+  - [CPP server README]() 
+  - [Python client README]()
+
+---
+
+### Screenshots
+#### Ex1 screenshots
+<details>
+<summary>Click to expand Ex1 screenshots</summary>
+<img src="/screenshots/BloomFilter/example%20run%201.png" height="500" alt="img1">
+<img src="/screenshots/BloomFilter/example%20run%202.png" height="500" alt="img2">
+<img src="/screenshots/BloomFilter/example%20run%203.png" height="500" alt="img3">
+<img src="/screenshots/BloomFilter/test%20results.png" height="500" alt="img4">
+</details>
+
+---
+
+#### Ex2 screenshots
+<details>
+<summary>Click to expand Ex2 screenshots</summary>
+TODO add
+</details>
+
+---
+
+## Useful links
+- [CPP server README]()
+- [Python client README]()
+
+---
