+++ server_cpp/CMakeLists.txt
@@ -16,25 +16,25 @@ FetchContent_MakeAvailable(googletest)
 
 # For compiling the project itself
 file(GLOB_RECURSE TEST_FILES
-src/utils/*.cpp
-src/bloom/*.cpp
-src/bloom/commands/*.cpp
-src/bloom/hash/*.cpp
-src/data_persistence/*.cpp
-tests/*.cpp
-tests/utils_tests/*.cpp
+        src/utils/*.cpp
+        src/bloom/*.cpp
+        src/bloom/commands/*.cpp
+        src/bloom/hash/*.cpp
+        src/data_persistence/*.cpp
+        tests/*.cpp
+        tests/utils_tests/*.cpp
 )
 # For compiling the tests
 add_executable(runTests ${TEST_FILES})
 
 # For compiling the project itself
 file(GLOB_RECURSE SRC_FILES
-src/*.cpp
-src/utils/*.cpp
-src/bloom/*.cpp
-src/bloom/commands/*.cpp
-src/bloom/hash/*.cpp
-src/data_persistence/*.cpp
+        src/*.cpp
+        src/utils/*.cpp
+        src/bloom/*.cpp
+        src/bloom/commands/*.cpp
+        src/bloom/hash/*.cpp
+        src/data_persistence/*.cpp
 )
 add_executable(app ${SRC_FILES})
