+++ server_cpp/Dockerfile
@@ -7,7 +7,7 @@ FROM gcc:latest
 RUN apt-get update && apt-get install -y cmake
 # Copy and set the project in the container
 WORKDIR /usr/src/app
-COPY . .
+COPY .. .
 
 # Ensures a clean build folder
 RUN rm -rf build && mkdir build
@@ -15,5 +15,5 @@ WORKDIR /usr/src/app/build
 
 # Compile the CPP project, add more files and flags as needed
 RUN cmake .. && make
-# Run the compiled project
+# Run the compiled project # TODO rename to server matching name
 CMD ["./app"]
